Add Copy and Parallel modifiers
The parallel modifier is like the sequential modifier, with the
difference that all the modifiers will be applied to the same function
and won't be chained. The result of the individual applications is
combined with the monoid append operator.

The copy modifier copies an entry in the schedule from one location to
another one. It does not remove the the entry at the old location, nor
does it preserve the entry at the new location. Wrap it in Copy to do that.