Support GHC-7.4 again
This commit adds a workaround for a bug in GHC 7.4 (I think it was
http://hackage.haskell.org/trac/ghc/ticket/5751), which previously
caused us to drop support for it. The bug was caused by the Functor
superclass constraint on Ixed/Contains from Lens, so for the fix there
are copies of these typeclasses without this constraint (New module:
Core.AtSansFunctor). You can then implement the ix/contains functions
for that typeclass and at an instance for the lens' Contains/Ixed in
terms of that instances and it'll work fine